Allow fetching secrets with federation disabled
When fetching a secret but passing "federation=False", disable support
for federation and force a rewrite of the secret href to the provided
local Barbican endpoint.
This allows for service usage when secrets are created with a public
endpoint but must be retrieved from an internal or admin endpoint.

Change-Id: I90778a2eeefc4cfe42b0e2a48ba09036e3e6d83d
Story: 2003197
Task: 23353